made variable long int

+++ 0x02-functions_nested_loops/7-print_last_digit.c
@@ -8,7 +8,7 @@
 
 int print_last_digit(int n)
 {
-	int x = (n % 10);
+	long int x = (n % 10);
 
 	_putchar(x + '0');
 	return (x);
